rename Tab so it does not generate a doc
